@@ -2,9 +2,9 @@ apply plugin: 'java'
 apply plugin: 'war'
 apply plugin: 'eclipse-wtp'
 
-// JDK 7
-sourceCompatibility = 13
-targetCompatibility = 13
+// JDK 17
+sourceCompatibility = 17
+targetCompatibility = 17
 
 repositories {
     mavenLocal()
@@ -14,30 +14,34 @@ repositories {
 
 dependencies {
  	
- 	// Use JUnit test framework.
+// Use JUnit test framework.
     testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.mockito:mockito-core:3.12.4'
+    testImplementation 'org.mockito:mockito-core:4.9.0'
  	implementation group: 'aopalliance', name: 'aopalliance', version: '1.0'
- 	implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.12.4'
+ 	implementation 'com.fasterxml.jackson.core:jackson-databind:2.14.1'
  	implementation 'commons-logging:commons-logging:1.2'
- 	implementation 'org.jsoup:jsoup:1.14.3'
-	implementation 'ch.qos.logback:logback-classic:1.2.11'
+ 	implementation 'org.jsoup:jsoup:1.15.3'
+	implementation 'ch.qos.logback:logback-classic:1.4.5'
 	//Spring
-	implementation 'org.springframework:spring-aop:5.3.18'
-	implementation 'org.springframework:spring-beans:5.3.18'
-	implementation 'org.springframework:spring-context:5.3.18'
-	implementation 'org.springframework:spring-core:5.3.18'
-	implementation 'org.springframework:spring-expression:5.3.18'
-	implementation 'org.springframework:spring-webmvc:5.3.18'
-	implementation 'org.springframework:spring-web:5.3.18'
-	implementation 'javax.servlet:jstl:1.2'
-	implementation group: 'javax.servlet', name: 'javax.servlet-api', version: '3.0.1'	
+	implementation 'org.springframework:spring-aop:6.0.0'
+	implementation 'org.springframework:spring-beans:6.0.0'
+	implementation 'org.springframework:spring-context:6.0.0'
+	implementation 'org.springframework:spring-core:6.0.0'
+	implementation 'org.springframework:spring-expression:6.0.0'
+	implementation 'org.springframework:spring-webmvc:6.0.0'
+	implementation 'org.springframework:spring-web:6.0.0'
+	implementation 'jakarta.el:jakarta.el-api:5.0.0'
+	implementation 'jakarta.servlet.jsp.jstl:jakarta.servlet.jsp.jstl-api:3.0.0'
+	compileOnly 'jakarta.servlet.jsp:jakarta.servlet.jsp-api:3.1.0'	
+	compileOnly 'jakarta.servlet:jakarta.servlet-api:6.0.0'
+	compileOnly 'jakarta.websocket:jakarta.websocket-api:2.1.0'
+	implementation 'org.glassfish.web:jakarta.servlet.jsp.jstl:3.0.1'
 	implementation 'mysql:mysql-connector-java:8.0.21'
-	implementation 'org.slf4j:slf4j-api:1.7.7'
+	implementation 'org.slf4j:slf4j-api:2.0.5'
 	// This dependency is used internally, and not exposed to consumers on their own compile classpath.
-    implementation 'com.google.guava:guava:29.0-jre'
-    implementation 'com.dropbox.core:dropbox-core-sdk:4.0.0'
-    implementation 'com.auth0:java-jwt:3.19.0'
+    implementation 'com.google.guava:guava:31.1-jre'
+    implementation 'com.dropbox.core:dropbox-core-sdk:5.4.4'
+    implementation 'com.auth0:java-jwt:4.2.1'
 	
 }
 
@@ -45,13 +49,17 @@ dependencies {
 eclipse {
 
   wtp {
+  
+  	facet {
+      file {
+      	facet name: 'jst.web', version: '5.0'
+      }
+    }
+    
     component {
-      
       //define context path, default to project folder name
-      contextPath = '/math'
-      
+      contextPath = 'math'
     }
     
   }
 }
-
